PRD-4964 - When inserting subreports and crosstabs, add the default "*" parameter to the import params. It is a sensible default. Make sure that both drag-n-drop and insert from the menu work the same in that respect.
