+++ designer/report-designer/src/org/pentaho/reporting/designer/core/editor/report/elements/AbstractSubReportElementDragHandler.java
@@ -70,4 +70,5 @@ public abstract class AbstractSubReportElementDragHandler extends AbstractReport
 
     return dragContext.getDefaultElement();
   }
+
 }
